requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
datasets==3.3.0
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.17.0
Flask==3.1.0
Flask-Limiter==3.5.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
limits==4.0.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
multiprocess==0.70.16
mysqlclient==2.2.7
numpy==2.2.3
openai==1.63.0
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
propcache==0.2.1
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pytest==7.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
supervisor==4.2.5
tenacity==9.0.0
tokenizers==0.21.0
tomli==2.2.1
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
